# Conflicts:
#	CMakeLists.txt
#	Common/ac/game_version.h
#	Common/ac/gamestructdefines.h
#	Common/core/def_version.h
#	Common/game/room_file.cpp
#	Common/game/roomstruct.cpp
#	Common/game/roomstruct.h
#	Common/gfx/allegrobitmap.cpp
#	Common/util/wgt2allg.cpp
#	Common/util/wgt2allg.h
#	Editor/AGS.Editor/Components/TranslationsComponent.cs
#	Editor/AGS.Editor/Panes/Room/RoomEditFilters/BaseAreasEditorFilter.cs
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Native/agsnative.cpp
#	Editor/AGS.Types/Enums/RoomResolution.cs
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Engine/ac/button.cpp
#	Engine/ac/character.cpp
#	Engine/ac/character.h
#	Engine/ac/dialog.cpp
#	Engine/ac/display.cpp
#	Engine/ac/draw.cpp
#	Engine/ac/dynamicsprite.cpp
#	Engine/ac/game.cpp
#	Engine/ac/global_character.cpp
#	Engine/ac/global_display.cpp
#	Engine/ac/invwindow.cpp
#	Engine/ac/movelist.cpp
#	Engine/ac/movelist.h
#	Engine/ac/object.cpp
#	Engine/ac/overlay.cpp
#	Engine/ac/room.cpp
#	Engine/ac/route_finder.cpp
#	Engine/ac/route_finder.h
#	Engine/ac/runtime_defines.h
#	Engine/ac/screenoverlay.cpp
#	Engine/game/savegame_components.cpp
#	Engine/gui/guidialog.cpp
#	Engine/media/audio/soundclip.cpp
#	Engine/media/audio/soundclip.h
#	Engine/plugin/agsplugin.cpp
#	Engine/script/scriptexecutor.cpp
#	version.json

+++ Engine/script/scriptexecutor.cpp
@@ -714,7 +714,7 @@ ScriptExecError ScriptExecutor::Run(int32_t curpc)
         codeOp.ArgCount = sccmd_info[codeOp.Instruction.Code].ArgCount;
 
         CC_ERROR_IF_RETCODE(static_cast<uint32_t>(_pc + codeOp.ArgCount) >= _codesize,
-            "unexpected end of code data (%d; %u)", _pc + codeOp.ArgCount, _codesize);
+            "unexpected end of code data (%u; %u)", static_cast<uint32_t>(_pc + codeOp.ArgCount), _codesize);
 
 
         // Read arguments; use switch as it proved to be faster than the loop
